agent-python/requirements.txt
pymongo
python-dotenv
openai
networkx
pytest
fastapi
uvicorn
httpx
matplotlib
pandas
numpy
